fix gcc not allowing logf

+++ fractal/mandelbrot.cpp
@@ -67,7 +67,7 @@ uint32_t* Mandelbrot::generate()
     // we distribute work by divisions for a better handling of unbalanced pictures
     int division_size = height_ / work_divisions;
     int from_y = 0;
-    for (int work = 0; work < thread_pool_->size(); ++work) {
+    for (int work = 0; work < static_cast<int>(thread_pool_->size()); ++work) {
         int to_y = from_y + (division_size - from_y) / (thread_pool_->size() - work);
         for (int division = 0; division < work_divisions; ++division)
             add_work(division * division_size + from_y, division * division_size + to_y);
@@ -94,8 +94,8 @@ uint32_t Mandelbrot::color(float x, float y, int iterations)
     if (iterations == max_iterations_)
         return palette_[max_iterations_];
 
-    float log_zn = std::logf(x * x + y * y) / 2;
-    float nu = std::logf(log_zn / std::logf(2)) / std::logf(2);
+    float log_zn = std::log(x * x + y * y) / 2;
+    float nu = std::log(log_zn / std::log(2)) / std::log(2);
 
     float smooth_iterations = iterations + 1 - nu;
     int smooth_iterations_int = static_cast<int>(smooth_iterations);
